[ENGG-2526] refactor: move request fetching from context to one level down (#2373)
* [ENGG-2485] fix: sidebar breadcrumb sync - navigate to root route (#2356)

* sidebar-breadcrumb sync,moving to index

* navigate to root route fix

* Revert "[ENGG-2485] fix: sidebar breadcrumb sync - navigate to root route (#2…" (#2365)

* refactor: move request fetching from context to one level down

---------
